add FaultTolerance.BulkheadBuilder.enableVirtualThreadsQueueing()
This method enables bulkhead queueing for synchronous actions executed
on virtual threads. The implementation is a simple two-semaphores bulkhead,
which has in fact already existed as `FutureThreadPoolBulkhead`. This commit
generalizes that class and leaves `FutureThreadPoolBulkhead` as a special
case, nearly empty, for guarding asynchronous actions returning `Future<V>`.
